refactor: restructure project layout and update imports for GreeMQTT package
