+++ GreeMQTT/managers.py
@@ -5,10 +5,10 @@ from typing import Callable, Tuple
 from loguru import logger
 import paho.mqtt.client as mqtt
 
-from device import search_devices, ScanResult
-from device_db import save_device
+from GreeMQTT.device import search_devices, ScanResult
+from GreeMQTT.device_db import save_device
 
-from mqtt_handler import handle_get_params, handle_set_params
+from GreeMQTT.mqtt_handler import handle_get_params, handle_set_params
 
 
 def run_thread(target: Callable, args: Tuple):
