+++ pyproject.toml
@@ -6,10 +6,13 @@ build-backend = "setuptools.build_meta"
 name = "GreeMQTT"
 version = "1.2.2"
 description = "MQTT bridge for Gree devices"
-readme = "README.md"
 authors = [
     { name = "monteship", email = "monteship@gmail.com" }
 ]
+readme = "README.md"
+license = { file = "LICENSE" }
+keywords = ["gree", "mqtt", "home-assistant"]
+
 requires-python = ">=3.12"
 dependencies = [
     "loguru~=0.7.3",
@@ -20,19 +23,26 @@ dependencies = [
     "cryptography~=44.0.3"
 ]
 classifiers = [
-    "Programming Language :: Python :: 3",
+    "Programming Language :: Python :: 3.12",
+    "Programming Language :: Python :: 3.13",
     "License :: OSI Approved :: MIT License",
-    "Operating System :: OS Independent"
+    "Operating System :: OS Independent",
+    "Topic :: Software Development :: Libraries :: Python Modules",
 ]
 
 [project.urls]
 Homepage = "https://github.com/monteship/GreeMQTT"
 
+[project.scripts]
+GreeMQTT = "GreeMQTT.__main__:main"
+
 [tool.setuptools]
-py-modules = ["config", "device", "device_db", "encryptor", "main", "managers", "mqtt_handler", "utils"]
+packages = ["GreeMQTT"]
 
 [tool.ruff]
-# Exclude a variety of commonly ignored directories.
+line-length = 88
+indent-width = 4
+target-version = "py313"
 exclude = [
     ".bzr",
     ".direnv",
@@ -62,47 +72,17 @@ exclude = [
     "venv",
 ]
 
-line-length = 88
-indent-width = 4
-target-version = "py313"
-
 [tool.ruff.lint]
-# Enable Pyflakes (`F`) and a subset of the pycodestyle (`E`)  codes by default.
-# Unlike Flake8, Ruff doesn't enable pycodestyle warnings (`W`) or
-# McCabe complexity (`C901`) by default.
 select = ["E4", "E7", "E9", "F"]
 ignore = []
-
-# Allow fix for all enabled rules (when `--fix`) is provided.
 fixable = ["ALL"]
 unfixable = []
-
-# Allow unused variables when underscore-prefixed.
 dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"
 
 [tool.ruff.format]
-# Like Black, use double quotes for strings.
 quote-style = "double"
-
-# Like Black, indent with spaces, rather than tabs.
 indent-style = "space"
-
-# Like Black, respect magic trailing commas.
 skip-magic-trailing-comma = false
-
-# Like Black, automatically detect the appropriate line ending.
 line-ending = "auto"
-
-# Enable auto-formatting of code examples in docstrings. Markdown,
-# reStructuredText code/literal blocks and doctests are all supported.
-#
-# This is currently disabled by default, but it is planned for this
-# to be opt-out in the future.
 docstring-code-format = false
-
-# Set the line length limit used when formatting code snippets in
-# docstrings.
-#
-# This only has an effect when the `docstring-code-format` setting is
-# enabled.
 docstring-code-line-length = "dynamic"
